+++ src/main/g8/app/grpc/GrpcApiUtils.java
@@ -1,44 +1,45 @@
-package thrift;
+package grpc;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.MissingNode;
 import com.fasterxml.jackson.databind.node.NullNode;
+import com.google.protobuf.ByteString;
 
-import thrift.def.TDataEncodingType;
+import grpc.def.ApiServiceProto.PDataEncodingType;
 import utils.ApiUtils;
 import utils.AppConstants;
 
 /**
- * Thrift API utility class.
+ * gRPC API utility class.
  * 
  * @author Thanh Nguyen <btnguyen2k@gmail.com>
- * @since template-v0.1.4
+ * @since template-v2.6.r2
  */
-public class ThriftApiUtils {
+public class GrpcApiUtils {
 
     /**
-     * Encode data from JSON to byte array.
+     * Encode data from JSON to {@link ByteString}.
      * 
      * @param dataType
      * @param jsonNode
      * @return
      */
-    public static byte[] encodeFromJson(TDataEncodingType dataType, JsonNode jsonNode) {
+    public static ByteString encodeFromJson(PDataEncodingType dataType, JsonNode jsonNode) {
         byte[] data = jsonNode == null || jsonNode instanceof NullNode
                 || jsonNode instanceof MissingNode ? null
                         : jsonNode.toString().getBytes(AppConstants.UTF8);
         if (data == null) {
-            return null;
+            return ByteString.EMPTY;
         }
         if (dataType == null) {
-            dataType = TDataEncodingType.JSON_STRING;
+            dataType = PDataEncodingType.JSON_STRING;
         }
         try {
             switch (dataType) {
             case JSON_STRING:
-                return data;
+                return ByteString.copyFrom(data);
             case JSON_GZIP:
-                return ApiUtils.toGzip(data);
+                return ByteString.copyFrom(ApiUtils.toGzip(data));
             default:
                 throw new IllegalArgumentException("Unsupported data encoding type: " + dataType);
             }
@@ -54,19 +55,19 @@ public class ThriftApiUtils {
      * @param data
      * @return
      */
-    public static JsonNode decodeToJson(TDataEncodingType dataType, byte[] data) {
-        if (data == null) {
-            return null;
+    public static JsonNode decodeToJson(PDataEncodingType dataType, ByteString data) {
+        if (data == null || data.isEmpty()) {
+            return NullNode.instance;
         }
         if (dataType == null) {
-            dataType = TDataEncodingType.JSON_STRING;
+            dataType = PDataEncodingType.JSON_STRING;
         }
         try {
             switch (dataType) {
             case JSON_STRING:
-                return ApiUtils.fromJsonString(data);
+                return ApiUtils.fromJsonString(data.toByteArray());
             case JSON_GZIP:
-                return ApiUtils.fromJsonGzip(data);
+                return ApiUtils.fromJsonGzip(data.toByteArray());
             default:
                 throw new IllegalArgumentException("Unsupported data encoding type: " + dataType);
             }